Very minor touches. Tested the four .c examples with 4 different compilers: VC 7.1 2003, gcc 7.4.0, mingw 7.3.0 and emcc (Emscripten gcc/clang-like replacement + linker emulating GNU ld) 1.38.40 clang version 10.0.0.

+++ c_hashtable.h
@@ -818,7 +818,7 @@ CH_API_DEF int CH_HASHTABLE_TYPE_FCT(_dbg_check)(const CH_HASHTABLE_TYPE* ht) {
     printf(". mem_used_percentage: %1.2f%% (100%% is the best possible result).\n",mem_used_percentage);
 #   endif
     CH_ASSERT(num_sorting_errors==0); /* When this happens, it can be a wrong user 'itemKey_cmp' function (that cannot sort keys in a consistent way) */
-    return num_total_items;
+    return (int) num_total_items;
 }
 
 CH_API_DEF void CH_HASHTABLE_TYPE_FCT(_swap)(CH_HASHTABLE_TYPE* a,CH_HASHTABLE_TYPE* b)  {
